improve files bundle readability
License: MIT
